Added LessonPair for multiple videos
