+++ src/index.css
@@ -21,3 +21,12 @@ a, h1, h2, p {
   color: white
 }
 
+.videoSelector {
+  width: 50%;
+  display: grid;
+  grid-template-columns: 85% 15%;
+}
+
+.addVideoButton {
+  margin: auto 0;
+}
